rest/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pika==1.2.0
redis==3.5.3
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
jsonpickle==2.0.0
Pillow==8.4.0
google-cloud-storage==1.43.0
mysql-connector==2.2.9
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0